refactor(cursor): hoist Kimi recovery into applyKimiToolCallRecovery helper
Collapses the two inline recovery blocks in cursor.ts into one-line calls to
a new applyKimiToolCallRecovery() in kimiToolCallNarration.ts. The helper
takes a structurally-typed ctx + optional emit callback (streaming) so no
executor import is needed and no module cycle forms.

This resolves the PR's Fast Quality Gate failures attributable to the change:
- file-size: cursor.ts was 1804 > frozen 1759; now 1763, baseline ratcheted
  to 1764 with justification (file hard-capped, base already 1758, hook is
  irreducible at 2 call sites — gate-sanctioned last resort).
- complexity-ratchets: cursor.ts 11→12 / 4→5 now OK (2762 vs baseline 3218).

Behavior unchanged — 9/9 kimi tests, 458/459 cursor+composer (1 skip).
Remaining CI failures (glm.ts TS2554, mutation quotaScoring, changelog
fragment) are pre-existing upstream regressions on release/v3.8.51.
